refacture with user

+++ src/controllers/wasteController.js
@@ -1,8 +1,14 @@
 const Waste = require('../models/wasteModel');
+const User = require('../models/userModel');
 
 // Create new waste
 const createWaste = async (req, res) => {
   try {
+    const userExists = await User.findById(req.body.userId);
+
+    if (!userExists) {
+      return res.status(404).send({ message: 'User tidak ditemukan' });
+    }
     const {
       jenis, berat, asalLimbah, harga, emisiKarbon,
     } = req.body;
@@ -25,6 +31,11 @@ const createWaste = async (req, res) => {
 // Get all wastes
 const getAllWastes = async (req, res) => {
   try {
+    const userExists = await User.findById(req.body.userId);
+
+    if (!userExists) {
+      return res.status(404).send({ message: 'User tidak ditemukan' });
+    }
     const wastes = await Waste.find();
     res.status(200).send(wastes);
   } catch (error) {
@@ -35,6 +46,11 @@ const getAllWastes = async (req, res) => {
 // Get a single waste by ID
 const getWasteById = async (req, res) => {
   try {
+    const userExists = await User.findById(req.body.userId);
+
+    if (!userExists) {
+      return res.status(404).send({ message: 'User tidak ditemukan' });
+    }
     const waste = await Waste.findById(req.params.id);
 
     if (!waste) {
@@ -50,6 +66,11 @@ const getWasteById = async (req, res) => {
 // Update a waste
 const updateWaste = async (req, res) => {
   try {
+    const userExists = await User.findById(req.body.userId);
+
+    if (!userExists) {
+      return res.status(404).send({ message: 'User tidak ditemukan' });
+    }
     const { id } = req.params;
     const {
       jenis, berat, asalLimbah, harga, emisiKarbon,
@@ -76,6 +97,11 @@ const updateWaste = async (req, res) => {
 // Delete a waste
 const deleteWaste = async (req, res) => {
   try {
+    const userExists = await User.findById(req.body.userId);
+
+    if (!userExists) {
+      return res.status(404).send({ message: 'User tidak ditemukan' });
+    }
     const { id } = req.params;
 
     const waste = await Waste.findByIdAndDelete(id);
